Move header to subfolder
