+++ shared/wallets/wallet/header/container.js
@@ -1,8 +1,8 @@
 // @flow
-import {connect, type TypedState} from '../../util/container'
-import * as Constants from '../../constants/wallets'
-import * as Types from '../../constants/types/wallets'
-import Header from './header'
+import {connect, type TypedState} from '../../../util/container'
+import * as Constants from '../../../constants/wallets'
+import * as Types from '../../../constants/types/wallets'
+import Header from '.'
 
 const mapStateToProps = (state: TypedState) => {
   const selectedAccount = Constants.getAccount(state)
