+++ shared/wallets/wallet/header/index.js
@@ -1,7 +1,7 @@
 // @flow
 import * as React from 'react'
-import * as Kb from '../../common-adapters'
-import * as Styles from '../../styles'
+import * as Kb from '../../../common-adapters'
+import * as Styles from '../../../styles'
 
 type Props = {
   isDefaultWallet: boolean,
